RE_LungCancerPrediction/requirements.txt
setuptools
Flask
joblib
Pillow
numpy
opencv-python
scikit-learn
gunicorn
